requirements.txt
praw==7.7.1
anthropic==0.18.1
yfinance==0.2.36
pandas==2.2.0
numpy==1.26.3
requests==2.31.0
python-dotenv==1.0.0